pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "raucle-detect"
version = "0.3.0"
description = "Open-source prompt injection detection for LLM applications"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Raucle", email = "oss@raucle.io" },
]
keywords = [
    "llm",
    "prompt-injection",
    "security",
    "ai-safety",
    "guardrails",
    "jailbreak-detection",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

# Zero mandatory dependencies -- the core scanner runs on the stdlib alone.
dependencies = []

[project.optional-dependencies]
rules = ["pyyaml>=6.0"]
ml = ["transformers>=4.30", "torch>=2.0"]
server = ["fastapi>=0.100", "uvicorn[standard]>=0.20", "pydantic>=2.0"]
dev = [
    "raucle-detect[rules,server]",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]
all = ["raucle-detect[rules,ml,server]"]

[project.scripts]
raucle-detect = "raucle_detect.cli:main"

[project.urls]
Homepage = "https://github.com/raucle/raucle-detect"
Documentation = "https://docs.raucle.io/raucle-detect"
Repository = "https://github.com/raucle/raucle-detect"
Issues = "https://github.com/raucle/raucle-detect/issues"

[tool.hatch.build.targets.wheel]
packages = ["raucle_detect"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
